requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.6
django-redis==5.4.0
djangorestframework==3.15.2
drf-extensions==0.7.1
drf-yasg==1.21.7
inflection==0.5.1
msgpack==1.1.0
packaging==24.2
Pillow==10.0.1
psycopg2==2.9.10
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
